# Python 3 compatibility