Hi, I do some changes in the code for run this package under Python 3. Check it please.
Greetings.