@erm 2015-02-09
I have been using this with Python 3.4 and Flask-SQLAlchemy for the past week and have experienced no problems.