@tonet666p 2015-02-09
erm try run "python3.4 setup.py test"