@erm 2015-02-10
Tests run fine.  Here is the output (my venv is Python 3.4):
    
    (venv) erm@localhost:~/Coding/dev/minibin/lib/Flask-WhooshAlchemy$ python setup.py test
    running test
    Searching for Flask-Testing
    Reading https://pypi.python.org/simple/Flask-Testing/
    Best match: Flask-Testing 0.4.2
    Downloading https://pypi.python.org/packages/source/F/Flask-Testing/Flask-Testing-0.4.2.tar.gz#md5=f47d9b859bae7ac43728cf61ca3eaa88
    Processing Flask-Testing-0.4.2.tar.gz
    Writing /tmp/easy_install-9szdl9f2/Flask-Testing-0.4.2/setup.cfg
    Running Flask-Testing-0.4.2/setup.py -q bdist_egg --dist-dir /tmp/easy_install-9szdl9f2/Flask-Testing-0.4.2/egg-dist-tmp-czqq5oej
    warning: no previously-included files matching '*.pyc' found under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'
    no previously-included directories found matching 'docs/_build'
    no previously-included directories found matching 'docs/_themes/.git'

    Installed /home/erm/Coding/dev/minibin/lib/Flask-WhooshAlchemy/Flask_Testing-0.4.2-py3.4.egg
    running egg_info
    writing top-level names to Flask_WhooshAlchemy.egg-info/top_level.txt
    writing requirements to Flask_WhooshAlchemy.egg-info/requires.txt
    writing Flask_WhooshAlchemy.egg-info/PKG-INFO
    writing dependency_links to Flask_WhooshAlchemy.egg-info/dependency_links.txt
    reading manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    reading manifest template 'MANIFEST.in'
    writing manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    running build_ext
    test_all (test.test_all.Tests) ... ok
    test_custom_analyzer (test.test_all.Tests) ... /home/erm/Coding/dev/minibin/lib/Flask-WhooshAlchemy/test/test_all.py:285: DeprecationWarning: Please use assertEqual instead.
      self.assertEquals(len(list(ObjectD.query.whoosh_search('trovel'))), 3)
    ok
    test_default_analyzer (test.test_all.Tests) ... ok
    test_invalid_attribute (test.test_all.Tests) ... ok

    ----------------------------------------------------------------------
    Ran 4 tests in 0.435s

    OK
    (venv) erm@localhost:~/Coding/dev/minibin/lib/Flask-WhooshAlchemy$