@tonet666p 2015-02-10
You are not using python3.4 command, i'm packaging it for fedora and it fails with python3X