@erm 2015-02-12
Ok, going to test the master branch using both Python 2.7 and 3.4:

    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ git branch
    * master
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ python --version
    Python 2.7.3
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ python setup.py test
    running test
    running egg_info
    writing requirements to Flask_WhooshAlchemy.egg-info/requires.txt
    writing Flask_WhooshAlchemy.egg-info/PKG-INFO
    writing top-level names to Flask_WhooshAlchemy.egg-info/top_level.txt
    writing dependency_links to Flask_WhooshAlchemy.egg-info/dependency_links.txt
    reading manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    reading manifest template 'MANIFEST.in'
    writing manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    running build_ext
    test_all (test.test_all.Tests) ... /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/SQLAlchemy-0.9.8-py2.7-linux-x86_64.egg/sqlalchemy/engine/default.py:573: SAWarning: Unicode type received non-unicodebind param value.
      param.append(processors[key](compiled_params[key]))
    ok
    test_custom_analyzer (test.test_all.Tests) ... ok
    test_default_analyzer (test.test_all.Tests) ... ok
    test_invalid_attribute (test.test_all.Tests) ... ok

    ----------------------------------------------------------------------
    Ran 4 tests in 0.302s

    OK
    Error in atexit._run_exitfuncs:
    Traceback (most recent call last):
      File "/usr/lib/python2.7/atexit.py", line 24, in _run_exitfuncs
        func(*targs, **kargs)
      File "/usr/lib/python2.7/multiprocessing/util.py", line 284, in _exit_function
        info('process shutting down')
    TypeError: 'NoneType' object is not callable
    Error in sys.exitfunc:
    Traceback (most recent call last):
      File "/usr/lib/python2.7/atexit.py", line 24, in _run_exitfuncs
        func(*targs, **kargs)
      File "/usr/lib/python2.7/multiprocessing/util.py", line 284, in _exit_function
        info('process shutting down')
    TypeError: 'NoneType' object is not callable

I fix this error in the test by importing the multprocessing module at the top of setup.py:

    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ python setup.py test
    running test
    running egg_info
    writing requirements to Flask_WhooshAlchemy.egg-info/requires.txt
    writing Flask_WhooshAlchemy.egg-info/PKG-INFO
    writing top-level names to Flask_WhooshAlchemy.egg-info/top_level.txt
    writing dependency_links to Flask_WhooshAlchemy.egg-info/dependency_links.txt
    reading manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    reading manifest template 'MANIFEST.in'
    writing manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    running build_ext
    test_all (test.test_all.Tests) ... /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/SQLAlchemy-0.9.8-py2.7-linux-x86_64.egg/sqlalchemy/engine/default.py:573: SAWarning: Unicode type received non-unicodebind param value.
      param.append(processors[key](compiled_params[key]))
    ok
    test_custom_analyzer (test.test_all.Tests) ... ok
    test_default_analyzer (test.test_all.Tests) ... ok
    test_invalid_attribute (test.test_all.Tests) ... ok

    ----------------------------------------------------------------------
    Ran 4 tests in 0.325s

    OK

So Python 2 works fine.  I will now test Python 3.4 on the master branch:

    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ /usr/local/bin/python3.4 setup.py test
    running test
    Searching for blinker==1.3
    Reading https://pypi.python.org/simple/blinker/
    Best match: blinker 1.3
    Downloading https://pypi.python.org/packages/source/b/blinker/blinker-1.3.tar.gz#md5=66e9688f2d287593a0e698cd8a5fbc57
    Processing blinker-1.3.tar.gz
    Writing /tmp/easy_install-r67nq1ci/blinker-1.3/setup.cfg
    Running blinker-1.3/setup.py -q bdist_egg --dist-dir /tmp/easy_install-r67nq1ci/blinker-1.3/egg-dist-tmp-i1lzelqz
    zip_safe flag not set; analyzing archive contents...

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/blinker-1.3-py3.4.egg
    Searching for Whoosh==2.6.0
    Reading https://pypi.python.org/simple/Whoosh/
    Best match: Whoosh 2.6.0
    Downloading https://pypi.python.org/packages/source/W/Whoosh/Whoosh-2.6.0.zip#md5=deb9be020263c20462e6bbb8053f6a0a
    Processing Whoosh-2.6.0.zip
    Writing /tmp/easy_install-ntxkn2le/Whoosh-2.6.0/setup.cfg
    Running Whoosh-2.6.0/setup.py -q bdist_egg --dist-dir /tmp/easy_install-ntxkn2le/Whoosh-2.6.0/egg-dist-tmp-l2dzy_d_
    warning: no files found matching '*.txt' under directory 'tests'
    warning: no files found matching '*.txt' under directory 'benchmark'
    warning: no files found matching '*.txt' under directory 'files'
    warning: no files found matching '*.py' under directory 'files'
    warning: no files found matching '*.jpg' under directory 'files'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Whoosh-2.6.0-py3.4.egg
    Searching for Flask-SQLAlchemy==1.0
    Reading https://pypi.python.org/simple/Flask-SQLAlchemy/
    Best match: Flask-SQLAlchemy 1.0
    Downloading https://pypi.python.org/packages/source/F/Flask-SQLAlchemy/Flask-SQLAlchemy-1.0.tar.gz#md5=dca034849ffcbaef7b848e8728d3668f
    Processing Flask-SQLAlchemy-1.0.tar.gz
    Writing /tmp/easy_install-x_kpr5pr/Flask-SQLAlchemy-1.0/setup.cfg
    Running Flask-SQLAlchemy-1.0/setup.py -q bdist_egg --dist-dir /tmp/easy_install-x_kpr5pr/Flask-SQLAlchemy-1.0/egg-dist-tmp-1ekya118
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'
    no previously-included directories found matching 'docs/_build'
    no previously-included directories found matching 'docs/_themes/.git'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Flask_SQLAlchemy-1.0-py3.4.egg
    Searching for Flask==0.10.1
    Reading https://pypi.python.org/simple/Flask/
    Best match: Flask 0.10.1
    Downloading https://pypi.python.org/packages/source/F/Flask/Flask-0.10.1.tar.gz#md5=378670fe456957eb3c27ddaef60b2b24
    Processing Flask-0.10.1.tar.gz
    Writing /tmp/easy_install-ulan_h2u/Flask-0.10.1/setup.cfg
    Running Flask-0.10.1/setup.py -q bdist_egg --dist-dir /tmp/easy_install-ulan_h2u/Flask-0.10.1/egg-dist-tmp-ihljy_7t
    warning: no files found matching '*' under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'
    warning: no previously-included files matching '*.pyc' found under directory 'tests'
    warning: no previously-included files matching '*.pyo' found under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'examples'
    warning: no previously-included files matching '*.pyo' found under directory 'examples'
    no previously-included directories found matching 'docs/_build'
    no previously-included directories found matching 'docs/_themes/.git'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Flask-0.10.1-py3.4.egg
    Searching for SQLAlchemy
    Reading https://pypi.python.org/simple/SQLAlchemy/
    Best match: SQLAlchemy 0.9.8
    Downloading https://pypi.python.org/packages/source/S/SQLAlchemy/SQLAlchemy-0.9.8.tar.gz#md5=470ca4da4a0081efc830f0d90dd91682
    Processing SQLAlchemy-0.9.8.tar.gz
    Writing /tmp/easy_install-_olqmsa9/SQLAlchemy-0.9.8/setup.cfg
    Running SQLAlchemy-0.9.8/setup.py -q bdist_egg --dist-dir /tmp/easy_install-_olqmsa9/SQLAlchemy-0.9.8/egg-dist-tmp-emmofivh
    warning: no files found matching '*.jpg' under directory 'doc'
    warning: no files found matching 'distribute_setup.py'
    warning: no files found matching 'sa2to3.py'
    warning: no files found matching 'ez_setup.py'
    no previously-included directories found matching 'doc/build/output'
    zip_safe flag not set; analyzing archive contents...
    sqlalchemy.testing.plugin.__pycache__.noseplugin.cpython-34: module references __file__
    sqlalchemy.__pycache__.cresultproxy.cpython-34: module references __file__
    sqlalchemy.__pycache__.cprocessors.cpython-34: module references __file__
    sqlalchemy.__pycache__.cutils.cpython-34: module references __file__

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/SQLAlchemy-0.9.8-py3.4-linux-x86_64.egg
    Searching for itsdangerous>=0.21
    Reading https://pypi.python.org/simple/itsdangerous/
    Best match: itsdangerous 0.24
    Downloading https://pypi.python.org/packages/source/i/itsdangerous/itsdangerous-0.24.tar.gz#md5=a3d55aa79369aef5345c036a8a26307f
    Processing itsdangerous-0.24.tar.gz
    Writing /tmp/easy_install-5sez4ba8/itsdangerous-0.24/setup.cfg
    Running itsdangerous-0.24/setup.py -q bdist_egg --dist-dir /tmp/easy_install-5sez4ba8/itsdangerous-0.24/egg-dist-tmp-heovijpt
    warning: no previously-included files matching '*' found under directory 'docs/_build'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/itsdangerous-0.24-py3.4.egg
    Searching for Jinja2>=2.4
    Reading https://pypi.python.org/simple/Jinja2/
    Best match: Jinja2 2.7.3
    Downloading https://pypi.python.org/packages/source/J/Jinja2/Jinja2-2.7.3.tar.gz#md5=b9dffd2f3b43d673802fe857c8445b1a
    Processing Jinja2-2.7.3.tar.gz
    Writing /tmp/easy_install-13m4lslp/Jinja2-2.7.3/setup.cfg
    Running Jinja2-2.7.3/setup.py -q bdist_egg --dist-dir /tmp/easy_install-13m4lslp/Jinja2-2.7.3/egg-dist-tmp-aionx1lt
    warning: no files found matching '*' under directory 'custom_fixers'
    warning: no previously-included files matching '*' found under directory 'docs/_build'
    warning: no previously-included files matching '*.pyc' found under directory 'jinja2'
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'jinja2'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Jinja2-2.7.3-py3.4.egg
    Searching for Werkzeug>=0.7
    Reading https://pypi.python.org/simple/Werkzeug/
    Best match: Werkzeug 0.10.1
    Downloading https://pypi.python.org/packages/source/W/Werkzeug/Werkzeug-0.10.1.tar.gz#md5=231d10996b7c3c1efe1776f22d570f83
    Processing Werkzeug-0.10.1.tar.gz
    Writing /tmp/easy_install-nraylxer/Werkzeug-0.10.1/setup.cfg
    Running Werkzeug-0.10.1/setup.py -q bdist_egg --dist-dir /tmp/easy_install-nraylxer/Werkzeug-0.10.1/egg-dist-tmp-q5bs5iup
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'
    warning: no previously-included files matching '*.pyc' found under directory 'tests'
    warning: no previously-included files matching '*.pyo' found under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'examples'
    warning: no previously-included files matching '*.pyo' found under directory 'examples'
    no previously-included directories found matching 'docs/_build'
    no previously-included directories found matching 'docs/_themes'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Werkzeug-0.10.1-py3.4.egg
    Searching for markupsafe
    Reading https://pypi.python.org/simple/markupsafe/
    Best match: MarkupSafe 0.23
    Downloading https://pypi.python.org/packages/source/M/MarkupSafe/MarkupSafe-0.23.tar.gz#md5=f5ab3deee4c37cd6a922fb81e730da6e
    Processing MarkupSafe-0.23.tar.gz
    Writing /tmp/easy_install-mbrvyie5/MarkupSafe-0.23/setup.cfg
    Running MarkupSafe-0.23/setup.py -q bdist_egg --dist-dir /tmp/easy_install-mbrvyie5/MarkupSafe-0.23/egg-dist-tmp-_41dsqmn

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/MarkupSafe-0.23-py3.4-linux-x86_64.egg
    Searching for Flask-Testing
    Reading https://pypi.python.org/simple/Flask-Testing/
    Best match: Flask-Testing 0.4.2
    Downloading https://pypi.python.org/packages/source/F/Flask-Testing/Flask-Testing-0.4.2.tar.gz#md5=f47d9b859bae7ac43728cf61ca3eaa88
    Processing Flask-Testing-0.4.2.tar.gz
    Writing /tmp/easy_install-u4j5wcfi/Flask-Testing-0.4.2/setup.cfg
    Running Flask-Testing-0.4.2/setup.py -q bdist_egg --dist-dir /tmp/easy_install-u4j5wcfi/Flask-Testing-0.4.2/egg-dist-tmp-sver_jzw
    warning: no previously-included files matching '*.pyc' found under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'tests'
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'
    no previously-included directories found matching 'docs/_build'
    no previously-included directories found matching 'docs/_themes/.git'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Flask_Testing-0.4.2-py3.4.egg
    running egg_info
    creating Flask_WhooshAlchemy.egg-info
    writing requirements to Flask_WhooshAlchemy.egg-info/requires.txt
    writing dependency_links to Flask_WhooshAlchemy.egg-info/dependency_links.txt
    writing top-level names to Flask_WhooshAlchemy.egg-info/top_level.txt
    writing Flask_WhooshAlchemy.egg-info/PKG-INFO
    writing manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    reading manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    reading manifest template 'MANIFEST.in'
    writing manifest file 'Flask_WhooshAlchemy.egg-info/SOURCES.txt'
    running build_ext
    Traceback (most recent call last):
      File "setup.py", line 40, in <module>
        test_suite='test.test_all',
      File "/usr/local/lib/python3.4/distutils/core.py", line 148, in setup
        dist.run_commands()
      File "/usr/local/lib/python3.4/distutils/dist.py", line 955, in run_commands
        self.run_command(cmd)
      File "/usr/local/lib/python3.4/distutils/dist.py", line 974, in run_command
        cmd_obj.run()
      File "/usr/local/lib/python3.4/site-packages/setuptools/command/test.py", line 138, in run
        self.with_project_on_sys_path(self.run_tests)
      File "/usr/local/lib/python3.4/site-packages/setuptools/command/test.py", line 118, in with_project_on_sys_path
        func()
      File "/usr/local/lib/python3.4/site-packages/setuptools/command/test.py", line 164, in run_tests
        testLoader = cks
      File "/usr/local/lib/python3.4/unittest/main.py", line 92, in __init__
        self.parseArgs(argv)
      File "/usr/local/lib/python3.4/unittest/main.py", line 139, in parseArgs
        self.createTests()
      File "/usr/local/lib/python3.4/unittest/main.py", line 146, in createTests
        self.module)
      File "/usr/local/lib/python3.4/unittest/loader.py", line 146, in loadTestsFromNames
        suites = [self.loadTestsFromName(name, module) for name in names]
      File "/usr/local/lib/python3.4/unittest/loader.py", line 146, in <listcomp>
        suites = [self.loadTestsFromName(name, module) for name in names]
      File "/usr/local/lib/python3.4/unittest/loader.py", line 105, in loadTestsFromName
        module = __import__('.'.join(parts_copy))
      File "/home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/test/test_all.py", line 88
        except OSError, e:
                      ^
    SyntaxError: invalid syntax

Why don't we just merge the Python 3 compatability branch into master at this point?:  

    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ git checkout python-3-compatibility 
    Branch python-3-compatibility set up to track remote branch python-3-compatibility from origin.
    Switched to a new branch 'python-3-compatibility'
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ ls
    doc  flask_whooshalchemy.py  LICENSE.txt  MANIFEST.in  README.rst  requirements.txt  setup.cfg  setup.py  test  tox.ini
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ git branch
      master
    * python-3-compatibility
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ git checkout master
    Switched to branch 'master'
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ git merge python-3-compatibility 
    Updating b0207cc..f24d729
    Fast-forward
     flask_whooshalchemy.py |    6 +++++-
     test/test_all.py       |    2 +-
     2 files changed, 6 insertions(+), 2 deletions(-)
    erm@localhost:~/Coding/dev/whooshtests/Flask-WhooshAlchemy$ python setup.py test
    running test
    Checking .pth file support in .
    /usr/bin/python -E -c pass
    Searching for blinker==1.3
    Reading http://pypi.python.org/simple/blinker/
    Best match: blinker 1.3
    Downloading https://pypi.python.org/packages/source/b/blinker/blinker-1.3.tar.gz#md5=66e9688f2d287593a0e698cd8a5fbc57
    Processing blinker-1.3.tar.gz
    Running blinker-1.3/setup.py -q bdist_egg --dist-dir /tmp/easy_install-ee47qg/blinker-1.3/egg-dist-tmp-vAzl_U
    zip_safe flag not set; analyzing archive contents...

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/blinker-1.3-py2.7.egg
    Searching for Whoosh==2.6.0
    Reading http://pypi.python.org/simple/Whoosh/
    Best match: Whoosh 2.6.0
    Downloading https://pypi.python.org/packages/source/W/Whoosh/Whoosh-2.6.0.zip#md5=deb9be020263c20462e6bbb8053f6a0a
    Processing Whoosh-2.6.0.zip
    Running Whoosh-2.6.0/setup.py -q bdist_egg --dist-dir /tmp/easy_install-CKaAjc/Whoosh-2.6.0/egg-dist-tmp-NRb4oO
    warning: no files found matching '*.txt' under directory 'tests'
    warning: no files found matching '*.txt' under directory 'benchmark'
    warning: no files found matching '*.txt' under directory 'files'
    warning: no files found matching '*.py' under directory 'files'
    warning: no files found matching '*.jpg' under directory 'files'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Whoosh-2.6.0-py2.7.egg
    Searching for Flask-SQLAlchemy==1.0
    Reading http://pypi.python.org/simple/Flask-SQLAlchemy/
    Best match: Flask-SQLAlchemy 1.0
    Downloading https://pypi.python.org/packages/source/F/Flask-SQLAlchemy/Flask-SQLAlchemy-1.0.tar.gz#md5=dca034849ffcbaef7b848e8728d3668f
    Processing Flask-SQLAlchemy-1.0.tar.gz
    Running Flask-SQLAlchemy-1.0/setup.py -q bdist_egg --dist-dir /tmp/easy_install-aAsf5a/Flask-SQLAlchemy-1.0/egg-dist-tmp-PZ8aqo
    warning: no previously-included files matching '*.pyc' found under directory 'docs'
    warning: no previously-included files matching '*.pyo' found under directory 'docs'
    no previously-included directories found matching 'docs/_build'
    no previously-included directories found matching 'docs/_themes/.git'

    Installed /home/erm/Coding/dev/whooshtests/Flask-WhooshAlchemy/Flask_SQLAlchemy-1.0-py2.7.egg
    Searching for SQLAlchemy
    Reading http://pypi.python.org/simple/SQLAlchemy/
    Best match: SQLAlchemy 0.9.8
    Downloading https://pypi.python.org/packages/source/S/SQLAlchemy/SQLAlchemy-0.9.8.tar.gz#md5=470ca4da4a0081efc830f0d90dd91682
    Processing SQLAlchemy-0.9.8.tar.gz
    Running SQLAlchemy-0.9.8/setup.py -q bdist_egg --dist-dir /tmp/easy_install-rRqcDZ/SQLAlchemy-0.9.8/egg-dist-tmp-_gq7vN
    warning: no files found matching '*.jpg' under directory 'doc'
    warning: no files found matching 'distribute_setup.py'
    warning: no files found matching 'sa2to3.py'
…